fix: address review feedback
